Changes to be committed: 	modified:   pfloat.vcp 	modified:   pinkyfloat.v
Added a fpu module that will serve as the floating point ALU. The int to
float conversion operation is currently being implemented. Need to learn
how to convert 16-bit int to 16-bit float using combinatorial logic.
fpu module var 'done' is the signal to the rest of the stages that an
ALU operation is currently running and no other processes should be
occurring during that time. Need to add logic to the other stages to
pause while !done.